Add /uploads endpoint that uses StaticFileHandler
iOS requires support for byte-range requests:

    https://developer.apple.com/library/archive/documentation/AppleApplications/Reference/SafariWebContent/CreatingVideoforSafarioniPhone/CreatingVideoforSafarioniPhone.html#//apple_ref/doc/uid/TP40006514-SW6

Our current raw handler does not support this, but the built-in
StaticFileHandler does.  Therefore, add a new /uploads endpoint and only
use for embedded video elements.

The reason why we don't simply use the StaticFileHandler for /raw is
because raw detects mime-type and handles extensions.